🧹 split scan.go into per-provider files with doc comments
Each scanner now lives in its own file (scan_ollama.go, scan_huggingface.go,
etc.) for easier review and navigation. Shared types, regex patterns, and
helper functions remain in scan.go.
